requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
jinja2==3.1.6
python-multipart==0.0.20
itsdangerous==2.2.0
Pillow==11.3.0
SQLAlchemy==2.0.43
psycopg2-binary==2.9.10
